pyproject.toml
[tool.poetry]
name = "django-graphene-m2m"
version = "0.1.0"
description = ""
authors = ["Fabien Le Frapper <contact@fabienlefrapper.me>"]

[tool.poetry.dependencies]
python = "^3.7"
graphene-django = "^2.12.1"
graphene = "^2.1.8"
django-silk = "^4.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
